添加竞赛类接口

+++ web/dao/mysqlDao/permission.go
@@ -11,10 +11,10 @@ func SelectPermissionByUid(ctx context.Context, uid string) (dao.Permission, err
 	err := db.Table(permission.TableName()).Where("uid=?", uid).Find(&permission).Error
 	return permission, err
 }
-func SelectPermissionList(ctx context.Context, offset, size int64) ([]dao.Permission, error) {
+func SelectPermissionList(ctx context.Context, offset, size int) ([]dao.Permission, error) {
 	db := GetDB(ctx)
 	ret := make([]dao.Permission, 0)
-	err := db.Table(dao.Permission{}.TableName()).Offset(int(offset)).Limit(int(size)).Find(&ret).Error
+	err := db.Table(dao.Permission{}.TableName()).Offset(offset).Limit(size).Find(&ret).Error
 	return ret, err
 }
